requirements.txt
streamlit
streamlit-option-menu
python-dotenv
requests
Pillow
webcolors
fal-client
